feat: put the new encoder in a separated from the original cli
