+++ README.md
@@ -19,6 +19,7 @@ A Bevy plugin for capturing frames from a Bevy application. It comes with some b
 | [`GifEncoder`](encoder::gif::GifEncoder)                              | Encodes frames into a gif.                                                | `gif`             |
 | [`Mp4Openh264Encoder`](encoder::mp4_openh264::Mp4Openh264Encoder)     | Encodes frames into an mp4 using openh264.                                | `mp4_openh264`    |
 | [`Mp4FfmpegCliEncoder`](encoder::mp4_ffmpeg_cli::Mp4FfmpegCliEncoder) | Encodes frames into an mp4 using the ffmpeg CLI (ffmpeg must be in PATH). | `mp4_ffmpeg_cli`  |
+| [`Mp4FfmpegCliPipeEncoder`](encoder::streaming_ffmpeg_cli_pipe::Mp4FfmpegCliPipeEncoder) | Encodes frames into an mp4 by streaming it directly to the ffmpeg CLI (ffmpeg must be in PATH). | `mp4_ffmpeg_cli_pipe`  |
 
 ## Usage
